Better Track Gear dialog presentation
